Rommel Scenario First Pass
First go at the Rommel Game's XML is here.  Also find a renamed "Notes" file for the Torch game, since I botched it the first time.
Once TGC gets me the setup chart for Torch I'll get that one plugged in too so we can get to the serious stuff.
Still waiting on my wife for the map, she's been extremely busy with work at the moment so a bit delayed.