Revision 0.20
Revision 0.20 - First draft at adding placement/unit-count limits to the various unit types.  This is incomplete as some of the existing "unit types" have unique placement/unit-count limits for the different powers that requires them to be split into power-specific "unit types".  Will get to this in a future update if someone else does not handle it.